explicit parsing of workshop cards and card parsing as submodule
