+++ src/routes/[[locale=locale]]/[education=education]/[resources=resources]/+page.svelte
@@ -2,27 +2,18 @@
   import {page_key} from '$lib/stores/page_key';
   import {onMount} from 'svelte';
   import WorkshopCard from '$lib/components/WorkshopCard.svelte';
+  import {parseEntries} from '$lib/js/parse_cms';
 
   onMount(() => {
     $page_key = 'navbar.education.resources';
   });
 
   export let data;
-  let workshops;
-  $: workshops = data.workshops;
+  $: workshops = parseEntries(data.workshops, 'workshops');
 </script>
 
 <div class="space-y-6">
-  {#each workshops as workshop, i}
-    <WorkshopCard
-      title={workshop.name}
-      subtitle={workshop.teaser}
-      language={workshop.language}
-      href={workshop.resource_link}
-      resp_unit={workshop.responsible_unit}
-      correlaidx_city={workshop.local_chapters[0]
-        ? workshop.local_chapters[0].Local_Chapters_id.translations[0].city
-        : null}
-    />
+  {#each workshops as workshop}
+    <WorkshopCard {...workshop} />
   {/each}
 </div>
